refactor project dump vitai
