+++ pipelines/rj_sms/__init__.py
@@ -6,5 +6,5 @@ Prefect flows for rj_sms project
 from pipelines.rj_sms.dump_db_sivep.flows import *
 from pipelines.rj_sms.pubsub.flows import *
 from pipelines.rj_sms.dump_cnes import *
-from pipelines.rj_sms.dump_vitai.flows import *
+from pipelines.rj_sms.dump_api_vitai.flows import *
 from pipelines.rj_sms.dump_azureblob_tpc.flows import *
